+++ platform-operator/controllers/verrazzano/component/common/crds_test.go
@@ -1,11 +1,18 @@
-// Copyright (c) 2022, Oracle and/or its affiliates.
+// Copyright (c) 2022, 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 package common
 
 import (
+	"fmt"
+	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
+	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	apiextv1fake "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/fake"
+	apiextv1 "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"testing"
 
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
@@ -62,4 +69,58 @@ func TestConvertVerrazzanoCR(t *testing.T) {
 	// THEN an appropriate error is returned
 	err = ConvertVerrazzanoCR(nil, &convertVZ)
 	assert.Error(t, err)
+}
+
+// TestCheckCRDsExist tests the CheckCRDsExist function
+// GIVEN a call to CheckCRDsExist
+// WHEN the requested CRDs are or aren't present
+// THEN true is returned if they are, false if not
+func TestCheckCRDsExist(t *testing.T) {
+	asserts := assert.New(t)
+
+	testCRDs := []string{
+		"foo",
+		"bar",
+	}
+	defer func() { k8sutil.ResetGetAPIExtV1ClientFunc() }()
+	k8sutil.GetAPIExtV1ClientFunc = func() (apiextv1.ApiextensionsV1Interface, error) {
+		return nil, fmt.Errorf("unexpected error")
+	}
+
+	exist, err := CheckCRDsExist(testCRDs)
+	asserts.Error(err)
+	asserts.False(exist)
+
+	k8sutil.GetAPIExtV1ClientFunc = func() (apiextv1.ApiextensionsV1Interface, error) {
+		return apiextv1fake.NewSimpleClientset().ApiextensionsV1(), nil
+	}
+
+	exist, err = CheckCRDsExist(testCRDs)
+	asserts.NoError(err)
+	asserts.False(exist)
+
+	k8sutil.GetAPIExtV1ClientFunc = func() (apiextv1.ApiextensionsV1Interface, error) {
+		return apiextv1fake.NewSimpleClientset(newTestCRDs(testCRDs...)...).ApiextensionsV1(), nil
+	}
+
+	exist, err = CheckCRDsExist(testCRDs)
+	asserts.NoError(err)
+	asserts.True(exist)
+}
+
+func newTestCRDs(crds ...string) []runtime.Object {
+	var runtimeObjs []runtime.Object
+	for _, crd := range crds {
+		runtimeObjs = append(runtimeObjs, newCRD(crd))
+	}
+	return runtimeObjs
+}
+
+func newCRD(name string) runtime.Object {
+	crd := &v1.CustomResourceDefinition{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: name,
+		},
+	}
+	return crd
 }