+++ tests/e2e/update/a-la-carte/a_la_carte_test.go
@@ -70,7 +70,7 @@ func (m appStackModifier) ModifyCR(cr *vzapi.Verrazzano) {
 
 	cr.Spec.Components.ApplicationOperator = &vzapi.ApplicationOperatorComponent{Enabled: &trueVal}
 	cr.Spec.Components.AuthProxy = &vzapi.AuthProxyComponent{Enabled: &trueVal}
-	cr.Spec.Components.CertManager = &vzapi.CertManagerComponent{Enabled: &trueVal}
+	cr.Spec.Components.CertManager = &vzapi.CertManagerComponent{Enabled: &falseVal}
 	cr.Spec.Components.Fluentd = &vzapi.FluentdComponent{Enabled: &trueVal}
 	cr.Spec.Components.Grafana = &vzapi.GrafanaComponent{Enabled: &trueVal}
 	cr.Spec.Components.Ingress = &vzapi.IngressNginxComponent{Enabled: &trueVal}