+++ tools/vz/cmd/install/install_test.go
@@ -394,47 +394,56 @@ func TestInstallCmdFilenamesAndSets(t *testing.T) {
 }
 
 // TestInstallCmdOperatorFile
-// GIVEN a CLI install command with defaults and --wait=false and --operator-file specified
+// GIVEN a CLI install command with defaults and --wait=false and --manifests OR the deprecated --operator-file is specified
 //
 //	WHEN I call cmd.Execute for install
 //	THEN the CLI install command is successful
-func TestInstallCmdOperatorFile(t *testing.T) {
-	c := fake.NewClientBuilder().WithScheme(helpers.NewScheme()).WithObjects(testhelpers.CreateTestVPOObjects()...).Build()
-	cmd, buf, errBuf, _ := createNewTestCommandAndBuffers(t, c)
-	cmd.PersistentFlags().Set(constants.OperatorFileFlag, "../../test/testdata/operator-file-fake.yaml")
-	cmd.PersistentFlags().Set(constants.WaitFlag, "false")
-	cmdHelpers.SetDeleteFunc(cmdHelpers.FakeDeleteFunc)
-	defer cmdHelpers.SetDefaultDeleteFunc()
+func TestInstallCmdManifestsFile(t *testing.T) {
+	tests := []struct {
+		testName          string
+		manifestsFlagName string
+	}{
+		{"Use manifests flag", constants.ManifestsFlag},
+		{"Use deprecated operator-file flag", constants.OperatorFileFlag},
+	}
+	for _, tt := range tests {
+		t.Run(tt.testName, func(t *testing.T) {
+			c := fake.NewClientBuilder().WithScheme(helpers.NewScheme()).WithObjects(testhelpers.CreateTestVPOObjects()...).Build()
+			cmd, buf, errBuf, _ := createNewTestCommandAndBuffers(t, c)
+			cmd.PersistentFlags().Set(tt.manifestsFlagName, "../../test/testdata/operator-file-fake.yaml")
+			cmd.PersistentFlags().Set(constants.WaitFlag, "false")
+			cmdHelpers.SetDeleteFunc(cmdHelpers.FakeDeleteFunc)
+			defer cmdHelpers.SetDefaultDeleteFunc()
+			cmdHelpers.SetVPOIsReadyFunc(func(_ client.Client) (bool, error) { return true, nil })
+			defer cmdHelpers.SetDefaultVPOIsReadyFunc()
+			// Run install command
+			err := cmd.Execute()
+			assert.NoError(t, err)
+			assert.Equal(t, "", errBuf.String())
+			assert.Contains(t, buf.String(), "Applying the file ../../test/testdata/operator-file-fake.yaml")
+			assert.Contains(t, buf.String(), "namespace/verrazzano-install created")
+			assert.Contains(t, buf.String(), "serviceaccount/verrazzano-platform-operator created")
+			assert.Contains(t, buf.String(), "service/verrazzano-platform-operator created\n")
 
-	cmdHelpers.SetVPOIsReadyFunc(func(_ client.Client) (bool, error) { return true, nil })
-	defer cmdHelpers.SetDefaultVPOIsReadyFunc()
+			// Verify the objects in the operator-file got added
+			sa := corev1.ServiceAccount{}
+			err = c.Get(context.TODO(), types.NamespacedName{Namespace: vzconstants.VerrazzanoInstallNamespace, Name: constants.VerrazzanoPlatformOperator}, &sa)
+			assert.NoError(t, err)
 
-	// Run install command
-	err := cmd.Execute()
-	assert.NoError(t, err)
-	assert.Equal(t, "", errBuf.String())
-	assert.Contains(t, buf.String(), "Applying the file ../../test/testdata/operator-file-fake.yaml")
-	assert.Contains(t, buf.String(), "namespace/verrazzano-install created")
-	assert.Contains(t, buf.String(), "serviceaccount/verrazzano-platform-operator created")
-	assert.Contains(t, buf.String(), "service/verrazzano-platform-operator created\n")
+			ns := corev1.Namespace{}
+			err = c.Get(context.TODO(), types.NamespacedName{Name: vzconstants.VerrazzanoInstallNamespace}, &ns)
+			assert.NoError(t, err)
 
-	// Verify the objects in the operator-file got added
-	sa := corev1.ServiceAccount{}
-	err = c.Get(context.TODO(), types.NamespacedName{Namespace: vzconstants.VerrazzanoInstallNamespace, Name: constants.VerrazzanoPlatformOperator}, &sa)
-	assert.NoError(t, err)
+			svc := corev1.Service{}
+			err = c.Get(context.TODO(), types.NamespacedName{Namespace: vzconstants.VerrazzanoInstallNamespace, Name: constants.VerrazzanoPlatformOperator}, &svc)
+			assert.NoError(t, err)
 
-	ns := corev1.Namespace{}
-	err = c.Get(context.TODO(), types.NamespacedName{Name: vzconstants.VerrazzanoInstallNamespace}, &ns)
-	assert.NoError(t, err)
-
-	svc := corev1.Service{}
-	err = c.Get(context.TODO(), types.NamespacedName{Namespace: vzconstants.VerrazzanoInstallNamespace, Name: constants.VerrazzanoPlatformOperator}, &svc)
-	assert.NoError(t, err)
-
-	// Verify the vz resource is as expected
-	vz := v1beta1.Verrazzano{}
-	err = c.Get(context.TODO(), types.NamespacedName{Namespace: "default", Name: "verrazzano"}, &vz)
-	assert.NoError(t, err)
+			// Verify the vz resource is as expected
+			vz := v1beta1.Verrazzano{}
+			err = c.Get(context.TODO(), types.NamespacedName{Namespace: "default", Name: "verrazzano"}, &vz)
+			assert.NoError(t, err)
+		})
+	}
 }
 
 // TestInstallValidations
@@ -445,17 +454,25 @@ func TestInstallCmdOperatorFile(t *testing.T) {
 func TestInstallValidations(t *testing.T) {
 	c := fake.NewClientBuilder().WithScheme(helpers.NewScheme()).WithObjects(testhelpers.CreateTestVPOObjects()...).Build()
 	cmd, _, _, _ := createNewTestCommandAndBuffers(t, c)
-	cmd.PersistentFlags().Set(constants.OperatorFileFlag, "test")
+	cmd.PersistentFlags().Set(constants.ManifestsFlag, "test")
 	cmd.PersistentFlags().Set(constants.VersionFlag, "test")
 	tempKubeConfigPath, _ := os.CreateTemp(os.TempDir(), testKubeConfig)
 	cmd.Flags().String(constants.GlobalFlagKubeConfig, tempKubeConfigPath.Name(), "")
 	cmd.Flags().String(constants.GlobalFlagContext, testK8sContext, "")
 	err := cmd.Execute()
 	assert.Error(t, err)
-	assert.Contains(t, err.Error(), fmt.Sprintf("--%s and --%s cannot both be specified", constants.VersionFlag, constants.OperatorFileFlag))
+	assert.Contains(t, err.Error(), fmt.Sprintf("--%s and --%s cannot both be specified", constants.VersionFlag, constants.ManifestsFlag))
 	if helpers.CheckAndRemoveBugReportExistsInDir("") {
 		t.Fatal(BugReportNotExist)
 	}
+
+	// test validation for deprecated operator-file flag and version being set
+	cmd, _, _, _ = createNewTestCommandAndBuffers(t, c)
+	cmd.PersistentFlags().Set(constants.OperatorFileFlag, "test")
+	cmd.PersistentFlags().Set(constants.VersionFlag, "test")
+	err = cmd.Execute()
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), fmt.Sprintf("--%s and --%s cannot both be specified", constants.VersionFlag, constants.ManifestsFlag))
 }
 
 // TestGetWaitTimeoutDefault