+++ CHANGES.md
@@ -3,7 +3,8 @@ Features:
 
 - The Prometheus components can now be enabled or disabled at any point in the Verrazzano lifecycle
 - Added a None profile that comes with all components disabled by default
-- Added Thanos, which supports high availability and long-term storage on top of Prometheus.
+- Added Thanos, which supports high availability and long-term storage on top of Prometheus
+- Enhanced the Verrazzano CLI to support installing and upgrading from a private registry and in air-gapped environments
 
 Component version updates:
 
@@ -11,6 +12,8 @@ Component version updates:
 - WebLogic Kubernetes Operator v4.0.6
 - WebLogic Monitoring Exporter v2.1.3
 - Jaeger v1.42.0
+- Prometheus Operator v0.64.1
+- kube-prometheus-stack Helm chart v45.25.0
 
 Components added:
 - Thanos v0.30.2